fixed the tests
