+++ tests/bdd/fixtures/failed_scenario_modal_fixtures.py
@@ -1,0 +1,14 @@
+import pytest
+from playwright.sync_api import Page
+
+@pytest.fixture
+def modal(page: Page):
+    return page.get_by_role("dialog")
+
+@pytest.fixture
+def modal_button(page: Page):
+    return page.locator("//*[contains(@class, 'failed-scenarios-link')]")
+
+@pytest.fixture
+def modal_close(page: Page):
+    return page.locator("//dialog//button[contains(@class, 'modal-close')]")
