+++ tests/bdd/steps/modal_steps.py
@@ -1,13 +1,33 @@
 from pytest_bdd import when, then, parsers
 from playwright.sync_api import Page, Locator, expect
+import pytest
+
+@pytest.fixture
+def modal_button(page: Page):
+    return page.locator("//*[contains(@class, 'failed-scenarios-link')]")
+
+@pytest.fixture
+def modal_close(page: Page):
+    return page.locator("//dialog//button[contains(@class, 'modal-close')]")
 
 @when("I open the failed scenarios modal")
-def open_modal(page: Page):
-    page.locator("//*[contains(@class, 'failed-scenarios-link')]").click()
-
+def open_modal(page: Page, modal_button: Locator):
+    modal_button.scroll_into_view_if_needed()
+    expect(modal_button).to_be_in_viewport()
+    modal_button.click()
+    
 @when("I close the modal")
-def close_modal(page: Page):
-    page.locator("//dialog//button[contains(@class, 'modal-close')]").click()
+def close_modal(page: Page, modal_close):
+    expect(modal_close).to_be_visible()
+    modal_close.click()
+    
+@then("the modal should be visible")
+def verify_modal_visible(page: Page, modal):
+    expect(modal).to_be_visible()
+
+@then("the modal should not be visible")
+def verify_modal_not_visible(page: Page):
+    expect(modal).not_to_be_visible()
 
 @then("the feature scenarios link should be visible")
 def feature_link_visible(page: Page):
