+++ tests/bdd/test_scenario_modal.py
@@ -1,5 +1,4 @@
 from pytest_bdd import scenarios
-# Importa gli step definiti altrove
 from tests.bdd.steps.setup_steps import *
 from tests.bdd.steps.common_steps import *
 from tests.bdd.steps.modal_steps import *
